break banner into small components
