--- app/components/UI/banner/Banner copy.js
@@ -1,11 +1,0 @@
-import BannerGallery from "@/components/UI/banner/BannerGallery";
-import BannerDescription from "@/components/UI/banner/FreeTourDescription";
-
-export default function Banner() {
-  return (
-    <div>
-      <BannerDescription />
-      <BannerGallery />
-    </div>
-  );
-}
