+++ app/components/UI/banner/Banner.js
@@ -1,0 +1,11 @@
+import BannerGallery from "./BannerGallery";
+import BannerDescription from "./FreeTourDescription";
+
+export default function Banner() {
+  return (
+    <div>
+      <BannerDescription />
+      <BannerGallery />
+    </div>
+  );
+}
