+++ app/components/UI/banner/BannerGallery.js
@@ -1,3 +1,6 @@
+import BannerImageCol from "./BannerImageCol";
+import PLACES from "./places";
+
 export default function BannerGallery() {
   return (
     <div className="mt-10">
@@ -7,61 +10,9 @@ export default function BannerGallery() {
       >
         <div className="absolute transform sm:left-1/2 sm:top-0 sm:translate-x-8 lg:left-1/2 lg:top-1/2 lg:-translate-y-1/2 lg:translate-x-8">
           <div className="flex items-center space-x-6 lg:space-x-8">
-            <div className="grid flex-shrink-0 grid-cols-1 gap-y-6 lg:gap-y-8">
-              <div className="h-64 w-44 overflow-hidden rounded-lg sm:opacity-0 lg:opacity-100">
-                <img
-                  src="https://images.unsplash.com/photo-1507496214531-ff4d0f45a58d?auto=format&fit=crop&q=80&w=1935&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D"
-                  alt=""
-                  className="h-full w-full object-cover object-center"
-                />
-              </div>
-              <div className="h-64 w-44 overflow-hidden rounded-lg">
-                <img
-                  src="https://images.unsplash.com/photo-1555881400-74d7acaacd8b?auto=format&fit=crop&q=80&w=1770&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D"
-                  alt=""
-                  className="h-full w-full object-cover object-center"
-                />
-              </div>
-            </div>
-            <div className="grid flex-shrink-0 grid-cols-1 gap-y-6 lg:gap-y-8">
-              <div className="h-64 w-44 overflow-hidden rounded-lg">
-                <img
-                  src="https://images.unsplash.com/photo-1648898975807-e2f2b59d7b61?auto=format&fit=crop&q=80&w=1936&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D"
-                  alt=""
-                  className="h-full w-full object-cover object-center"
-                />
-              </div>
-              <div className="h-64 w-44 overflow-hidden rounded-lg">
-                <img
-                  src="https://images.unsplash.com/photo-1562195168-c82fea0f0953?auto=format&fit=crop&q=80&w=1887&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D"
-                  alt=""
-                  className="h-full w-full object-cover object-center"
-                />
-              </div>
-              <div className="h-64 w-44 overflow-hidden rounded-lg">
-                <img
-                  src="https://images.unsplash.com/photo-1536663815808-535e2280d2c2?auto=format&fit=crop&q=80&w=1964&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D"
-                  alt=""
-                  className="h-full w-full object-cover object-center"
-                />
-              </div>
-            </div>
-            <div className="grid flex-shrink-0 grid-cols-1 gap-y-6 lg:gap-y-8">
-              <div className="h-64 w-44 overflow-hidden rounded-lg">
-                <img
-                  src="https://images.unsplash.com/photo-1496694971191-58e2450fc841?auto=format&fit=crop&q=80&w=1965&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D"
-                  alt=""
-                  className="h-full w-full object-cover object-center"
-                />
-              </div>
-              <div className="h-64 w-44 overflow-hidden rounded-lg">
-                <img
-                  src="https://images.unsplash.com/photo-1511968822213-92de73315bba?auto=format&fit=crop&q=80&w=1887&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D"
-                  alt=""
-                  className="h-full w-full object-cover object-center"
-                />
-              </div>
-            </div>
+            {PLACES.map((col) => (
+              <BannerImageCol images={col} />
+            ))}
           </div>
         </div>
       </div>
