+++ app/components/UI/banner/places.js
@@ -1,50 +1,56 @@
 const PLACES = [
-  {
-    id: "place1",
-    imgURL:
-      "https://images.unsplash.com/photo-1507496214531-ff4d0f45a58d?auto=format&fit=crop&q=80&w=1935&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
-    title: "Lisbon",
-  },
-  {
-    id: "place2",
-    imgURL:
-      "https://images.unsplash.com/photo-1555881400-74d7acaacd8b?auto=format&fit=crop&q=80&w=1770&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
-    title: "Porto",
-  },
-  {
-    id: "place3",
-    imgURL:
-      "https://images.unsplash.com/photo-1648898975807-e2f2b59d7b61?auto=format&fit=crop&q=80&w=1936&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
-    title: "Geres",
-  },
-  {
-    id: "place4",
-    imgURL:
-      "https://images.unsplash.com/photo-1562195168-c82fea0f0953?auto=format&fit=crop&q=80&w=1887&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
+  [
+    {
+      id: "place1",
+      imgUrl:
+        "https://images.unsplash.com/photo-1507496214531-ff4d0f45a58d?auto=format&fit=crop&q=80&w=1935&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
+      title: "Lisbon",
+    },
+    {
+      id: "place2",
+      imgUrl:
+        "https://images.unsplash.com/photo-1555881400-74d7acaacd8b?auto=format&fit=crop&q=80&w=1770&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
+      title: "Porto",
+    },
+  ],
+  [
+    {
+      id: "place3",
+      imgUrl:
+        "https://images.unsplash.com/photo-1648898975807-e2f2b59d7b61?auto=format&fit=crop&q=80&w=1936&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
+      title: "Geres",
+    },
+    {
+      id: "place4",
+      imgUrl:
+        "https://images.unsplash.com/photo-1562195168-c82fea0f0953?auto=format&fit=crop&q=80&w=1887&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
 
-    title: "Sintra",
-  },
-  {
-    id: "place5",
-    imgURL:
-      "https://images.unsplash.com/photo-1536663815808-535e2280d2c2?auto=format&fit=crop&q=80&w=1964&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
+      title: "Sintra",
+    },
+    {
+      id: "place5",
+      imgUrl:
+        "https://images.unsplash.com/photo-1536663815808-535e2280d2c2?auto=format&fit=crop&q=80&w=1964&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
 
-    title: "Lisbon",
-  },
-  {
-    id: "place6",
-    imgURL:
-      "https://images.unsplash.com/photo-1496694971191-58e2450fc841?auto=format&fit=crop&q=80&w=1965&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
+      title: "Lisbon",
+    },
+  ],
+  [
+    {
+      id: "place6",
+      imgUrl:
+        "https://images.unsplash.com/photo-1496694971191-58e2450fc841?auto=format&fit=crop&q=80&w=1965&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
 
-    title: "Algarve",
-  },
-  {
-    id: "place7",
-    imgURL:
-      "https://images.unsplash.com/photo-1511968822213-92de73315bba?auto=format&fit=crop&q=80&w=1887&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
+      title: "Algarve",
+    },
+    {
+      id: "place7",
+      imgUrl:
+        "https://images.unsplash.com/photo-1511968822213-92de73315bba?auto=format&fit=crop&q=80&w=1887&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
 
-    title: "Porto",
-  },
+      title: "Porto",
+    },
+  ],
 ];
 
 export default PLACES;
